+++ spec/lib/ember_cli/deploy/redis_spec.rb
@@ -2,53 +2,13 @@ require "spec_helper"
 require "ember_cli/deploy/redis"
 
 describe EmberCli::Deploy::Redis do
-  describe "#append_to_head" do
-    it "injects the string into the <head> tag" do
-      provided_html = "<html><head><title></title></head></html>"
-      ember_cli_deploy = build_ember_cli_deploy(index_html: provided_html)
-
-      ember_cli_deploy.append_to_head("<script></script>")
-      index_html = ember_cli_deploy.html
-
-      expect(index_html).to eq(
-        "<html><head><title></title><script></script></head></html>"
-      )
-    end
-  end
-
-  describe "#append_to_body" do
-    it "injects the string into the <body> tag" do
-      provided_html = "<html><body><br/></body></html>"
-      ember_cli_deploy = build_ember_cli_deploy(index_html: provided_html)
-
-      ember_cli_deploy.append_to_body("<script></script>")
-      ember_cli_deploy.append_to_body("<meta></meta>")
-      index_html = ember_cli_deploy.html
-
-      expect(index_html).to eq(
-        "<html><body><br/><script></script><meta></meta></body></html>"
-      )
-    end
-  end
-
-  describe "#html" do
-    context "when the HTML is provided" do
-      it "returns the HTML" do
-        provided_html = "<p>Hello World</p>"
-        ember_cli_deploy = build_ember_cli_deploy(index_html: provided_html)
-
-        index_html = ember_cli_deploy.html
-
-        expect(index_html).to eq(provided_html)
-      end
-    end
-
+  describe "#index_html" do
     context "when the keys are present" do
       it "retrieves the HTML from Redis" do
         stub_index_html(html: "<p>Hello World</p>")
         ember_cli_deploy = build_ember_cli_deploy
 
-        index_html = ember_cli_deploy.html
+        index_html = ember_cli_deploy.index_html
 
         expect(index_html).to eq("<p>Hello World</p>")
       end
@@ -60,7 +20,7 @@ describe EmberCli::Deploy::Redis do
         stub_index_html(html: nil, deploy_key: deploy_key)
         ember_cli_deploy = build_ember_cli_deploy
 
-        expect { ember_cli_deploy.html }.to raise_error(
+        expect { ember_cli_deploy.index_html }.to raise_error(
           /HTML for #{namespace}:#{deploy_key} is missing/
         )
       end
@@ -71,7 +31,7 @@ describe EmberCli::Deploy::Redis do
         stub_current_key(nil)
         ember_cli_deploy = build_ember_cli_deploy
 
-        expect { ember_cli_deploy.html }.to raise_error(
+        expect { ember_cli_deploy.index_html }.to raise_error(
           /#{namespace}:current is empty/
         )
       end
@@ -84,12 +44,13 @@ describe EmberCli::Deploy::Redis do
     end
   end
 
-  def build_ember_cli_deploy(index_html: nil)
-    EmberCli::Deploy::Redis.new(namespace: namespace, index_html: index_html)
+  def build_ember_cli_deploy
+    app = double(name: namespace)
+    EmberCli::Deploy::Redis.new(app)
   end
 
   def namespace
-    "human-health"
+    "ember-cli-rails-deploy-redis"
   end
 
   def stub_current_key(deploy_key)