+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 master
 ------
 
+* Integrate with `ember-cli-rails` deployment strategies. [#4]
+* Rename `EmberCLI` module to `EmberCli`. [#4]
+* Rename `ember-cli` directories to `ember_cli`. [#4]
+
+[#4]: https://github.com/seanpdoyle/ember-cli-rails-deploy-redis/pull/4
+
 0.0.2
 -----